+++ neurolabi/gui/zstackobjectsourcefactory.h
@@ -11,8 +11,8 @@ public:
 public:
   static std::string MakeWatershedBoundarySource(int label);
   static std::string MakeRectRoiSource(const std::string &suffix = "");
-  static std::string MakeFlyEmBodyMaskSource(int bodyId);
-  static std::string MakeFlyEmBodySource(int bodyId);
+  static std::string MakeFlyEmBodyMaskSource(uint64_t bodyId);
+  static std::string MakeFlyEmBodySource(uint64_t bodyId);
   static std::string MakeCurrentMsTileSource(int resLevel);
   static std::string MakeDvidTileSource();
   static std::string MakeDvidLabelSliceSource();
